:recycle: refactor: `ClovaSpeechClient` 인터페이스 분리, Inner Class 외부로 이동
