+++ src/main/java/com/mallang/mallang_backend/global/util/Boosting.java
@@ -1,0 +1,10 @@
+package com.mallang.mallang_backend.global.util;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class Boosting {
+	private String words;
+}
