+++ src/main/java/com/mallang/mallang_backend/global/util/Diarization.java
@@ -1,0 +1,12 @@
+package com.mallang.mallang_backend.global.util;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class Diarization {
+	private Boolean enable = Boolean.FALSE;
+	private Integer speakerCountMin;
+	private Integer speakerCountMax;
+}
